[#31] Move tokenizer to another folder
